requirements.txt
Django>=4.2,<5.0
django-crispy-forms>=2.0
django-import-export>=3.1
django-simple-history>=3.0
django-unfold>=0.18
python-dotenv>=1.0